refactor: 1ファイル1クラス構造へリファクタリングしパッケージ構造を導入
単一のmain.pyファイルに含まれていた5つのクラスと1つのデータクラスを、
役割ごとに適切なパッケージに分割して再構成した。

主な変更点:
- src/models: ImageMetricsデータクラスとGenreWeightsクラスを配置
- src/analyzers: MetricNormalizerとImageQualityAnalyzerクラスを配置
- src/services: GameScreenPickerクラスを配置
- src/main.py: CLIエントリーポイントのみに集約
- 各パッケージに__init__.pyを追加し、適切なエクスポート設定

これにより、コードの可読性と保守性が向上し、各クラスの責務が明確になった。
CLIの機能やインターフェースに変更はない。
