+++ src/__init__.py
@@ -1,0 +1,15 @@
+"""
+game_screen_pick - 高精度・コンテンツ多様性重視選択ツール
+"""
+
+from .models import ImageMetrics, GenreWeights
+from .analyzers import MetricNormalizer, ImageQualityAnalyzer
+from .services import GameScreenPicker
+
+__all__ = [
+    "ImageMetrics",
+    "GenreWeights",
+    "MetricNormalizer",
+    "ImageQualityAnalyzer",
+    "GameScreenPicker",
+]
