+++ src/analyzers/metric_normalizer.py
@@ -1,0 +1,30 @@
+"""
+Metric normalizer for image quality metrics.
+"""
+
+import math
+from typing import Dict
+
+
+class MetricNormalizer:
+    """メトリクス正規化クラス"""
+
+    @staticmethod
+    def sigmoid(x: float, center: float, steepness: float = 0.1) -> float:
+        try:
+            return 1 / (1 + math.exp(-steepness * (x - center)))
+        except:
+            return 1.0 if x > center else 0.0
+
+    @classmethod
+    def normalize_all(cls, raw: Dict[str, float]) -> Dict[str, float]:
+        return {
+            "blur_score": cls.sigmoid(raw['blur_score'], center=500, steepness=0.005),
+            "contrast": cls.sigmoid(raw['contrast'], center=50, steepness=0.1),
+            "color_richness": cls.sigmoid(raw['color_richness'], center=40, steepness=0.1),
+            "edge_density": min(1.0, raw['edge_density'] * 5.0),
+            "dramatic_score": min(1.0, raw['dramatic_score'] / 100.0),
+            "visual_balance": raw['visual_balance'] / 100.0,
+            "action_intensity": cls.sigmoid(raw['action_intensity'], center=30, steepness=0.2),
+            "ui_density": cls.sigmoid(raw['ui_density'], center=10, steepness=0.3)
+        }
