+++ src/models/__init__.py
@@ -1,0 +1,6 @@
+"""Models package for data classes and weight definitions."""
+
+from .image_metrics import ImageMetrics
+from .genre_weights import GenreWeights
+
+__all__ = ["ImageMetrics", "GenreWeights"]
