+++ src/services/__init__.py
@@ -1,0 +1,5 @@
+"""Services package for game screen selection logic."""
+
+from .screen_picker import GameScreenPicker
+
+__all__ = ["GameScreenPicker"]
